+++ src/Products/MeetingCommunes/adapters.py
@@ -658,8 +658,7 @@ class CustomMeetingConfig(MeetingConfig):
                         'sort_on': u'created',
                         'sort_reversed': True,
                         'showNumberOfItems': False,
-                        'tal_condition': "python: not tool.isPowerObserverForCfg(cfg) and "
-                            "not tool.isPowerObserverForCfg(cfg, isRestricted=True)",
+                        'tal_condition': "python: not tool.isPowerObserverForCfg(cfg)",
                         'roles_bypassing_talcondition': ['Manager', ]
                     }
                  ),