+++ CHANGES.rst
@@ -13,6 +13,7 @@ The Products.MeetingCommunes version must be the same as the Products.PloneMeeti
 - In profile 'examples_fr', enable 'groups_in_charge' for 'Secrétariat Général' and configure auto asked advice for it
 - In profile 'examples_fr', enable 'MeetingItem.manuallyLinkedItems' field
 - In profile 'examples_fr', enable 'Agenda with annexes' by default
+- Adapted code regarding MeetingConfig.powerObservers
 
 4.1b2 (2019-01-29)
 ------------------